fix: resolve all clippy warnings across workspace
- Replace redundant closures with function references in grpc.rs
- Use is_empty() instead of len() > 0 / len() == 0
- Remove unnecessary mut qualifiers on IVF-PQ index variables
- Replace struct field reassignment with initialization syntax
- Fix approximate constant in tests (3.14 → 3.15)
- Remove redundant type casts (u64 as u64)
- Replace vec![] with array literals in benchmark loops

+++ examples/basic_usage.rs
@@ -40,7 +40,7 @@ fn main() {
     // Quick throughput estimate
     let num_vectors = 10_000;
     let vectors: Vec<Vector> = (0..num_vectors).map(|i| Vector::random(i, dim)).collect();
-    let query = Vector::random(num_vectors as u64, dim);
+    let query = Vector::random(num_vectors, dim);
 
     let start = std::time::Instant::now();
     let _sum: f32 = vectors
